requirements.txt
asgiref==3.5.0
Django==3.0
djangorestframework==3.11.0
psycopg2==2.8
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
